Start project over
Created function to return value of 'paper', 'rock', or scissors.
Created a function to prompt the user to input 'paper', 'rock', or 'scissors'
Changed the input so that the user input would always enter as lowercase.
Created a function that would print to the console if the user or computer wins the round.